feat(QTDI-1305): rename sample supporterror to entry-with-error.
