+++ sample-parent/sample-features/entry-with-error/pom.xml
@@ -18,70 +18,49 @@
 
   <parent>
     <groupId>org.talend.sdk.component</groupId>
-    <artifactId>sample-parent</artifactId>
+    <artifactId>sample-features</artifactId>
     <version>1.81.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>sample-features</artifactId>
-  <packaging>pom</packaging>
+  <groupId>org.talend.sdk.component.sample.feature</groupId>
+  <artifactId>supporterror</artifactId>
+  <packaging>jar</packaging>
 
-  <name>Component Runtime :: Sample Features</name>
-
-  <modules>
-    <module>fixed-schema</module>
-    <module>aftergroup-lastgroup</module>
-    <module>conditional-outputs</module>
-    <module>checkpoint-runner</module>
-    <module>entry-with-error</module>
-  </modules>
+  <name>Component Runtime :: Sample Feature :: Entry with error support</name>
 
   <dependencies>
     <dependency>
       <groupId>org.talend.sdk.component</groupId>
       <artifactId>component-api</artifactId>
       <version>${project.version}</version>
-      <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.talend.sdk.component</groupId>
+      <artifactId>component-runtime-manager</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.tomitribe</groupId>
+      <artifactId>tomitribe-crest</artifactId>
+      <version>0.32</version>
     </dependency>
   </dependencies>
 
   <build>
     <plugins>
       <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <configuration>
+          <source>1.8</source>
+          <target>1.8</target>
+        </configuration>
+      </plugin>
+      <plugin>
         <groupId>org.talend.sdk.component</groupId>
         <artifactId>talend-component-maven-plugin</artifactId>
         <version>${project.version}</version>
-        <!-- Please look at the TCK documentation to understand the talend-component-maven-plugin -->
-        <!-- https://talend.github.io/component-runtime/main/latest/build-tools-maven.html -->
         <executions>
-          <execution>
-            <id>talend-dependencies</id>
-            <goals>
-              <goal>dependencies</goal>
-            </goals>
-            <phase>process-classes</phase>
-          </execution>
-          <execution>
-            <id>talend-component-bundle</id>
-            <goals>
-              <goal>car</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <excludeArtifactsFilter>com.fasterxml.jackson.core:jackson-databind,com.fasterxml.jackson.core:jackson-core,com.fasterxml.jackson.core:jackson-annotations</excludeArtifactsFilter>
-            </configuration>
-          </execution>
-          <execution>
-            <id>talend-scan-descriptor</id>
-            <goals>
-              <goal>scan-descriptor</goal>
-            </goals>
-          </execution>
-          <execution>
-            <id>talend-svg2png</id>
-            <goals>
-              <goal>svg2png</goal>
-            </goals>
-          </execution>
           <execution>
             <id>talend-component-validate</id>
             <goals>
@@ -98,8 +77,10 @@
               <validateModel>true</validateModel>
               <validateMetadata>true</validateMetadata>
               <validateComponent>true</validateComponent>
-              <validateDataStore>true</validateDataStore>
-              <validateDataSet>true</validateDataSet>
+              <validateDataStore>false</validateDataStore>
+              <!-- Don't want datastore... -->
+              <validateDataSet>false</validateDataSet>
+              <!-- ...nor dataset -->
               <validateActions>true</validateActions>
               <validateDocumentation>true</validateDocumentation>
               <validateWording>true</validateWording>
@@ -115,6 +96,46 @@
             </configuration>
           </execution>
         </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-jar-plugin</artifactId>
+        <configuration>
+          <archive>
+            <manifest>
+              <addClasspath>true</addClasspath>
+              <mainClass>org.talend.sdk.component.sample.feature.entrywitherror.Cli</mainClass>
+            </manifest>
+          </archive>
+        </configuration>
+      </plugin>
+      <plugin>
+        <artifactId>maven-shade-plugin</artifactId>
+        <version>2.1</version>
+        <executions>
+          <execution>
+            <goals>
+              <goal>shade</goal>
+            </goals>
+            <phase>package</phase>
+            <configuration>
+              <transformers>
+                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                  <mainClass>org.talend.sdk.component.sample.feature.entrywitherror.Cli</mainClass>
+                </transformer>
+              </transformers>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <version>3.0.0</version>
+        <configuration>
+          <mainClass>org.talend.sdk.component.sample.feature.entrywitherror.Cli</mainClass>
+        </configuration>
       </plugin>
     </plugins>
   </build>
