+++ sample-parent/sample-features/entry-with-error/src/main/java-templates/org/talend/sdk/component/sample/feature/entrywitherror/Versions.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright (C) 2006-2025 Talend Inc. - www.talend.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.talend.sdk.component.sample.feature.entrywitherror;
+
+public interface Versions {
+
+    String KIT_VERSION = "${project.version}";
+
+}
