+++ sample-parent/sample-features/entry-with-error/src/main/resources/META-INF/services/org.talend.sdk.component.runtime.serialization.ContainerFinder
