--- sample-parent/sample-features/supporterror/pom.xml
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2006-2025 Talend Inc. - www.talend.com
-   Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.talend.sdk.component</groupId>
-    <artifactId>sample-features</artifactId>
-    <version>1.81.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.talend.sdk.component.sample.feature</groupId>
-  <artifactId>supporterror</artifactId>
-  <packaging>jar</packaging>
-
-  <name>Component Runtime :: Sample Feature :: Supporterror</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.talend.sdk.component</groupId>
-      <artifactId>component-api</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.talend.sdk.component</groupId>
-      <artifactId>component-runtime-manager</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.tomitribe</groupId>
-      <artifactId>tomitribe-crest</artifactId>
-      <version>0.32</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.talend.sdk.component</groupId>
-        <artifactId>talend-component-maven-plugin</artifactId>
-        <version>${project.version}</version>
-        <executions>
-          <execution>
-            <id>talend-component-validate</id>
-            <goals>
-              <goal>validate</goal>
-            </goals>
-            <phase>process-classes</phase>
-            <configuration>
-              <validatePlaceholder>true</validatePlaceholder>
-              <validateSvg>true</validateSvg>
-              <validateLegacyIcons>false</validateLegacyIcons>
-              <validateFamily>true</validateFamily>
-              <validateSerializable>true</validateSerializable>
-              <validateInternationalization>true</validateInternationalization>
-              <validateModel>true</validateModel>
-              <validateMetadata>true</validateMetadata>
-              <validateComponent>true</validateComponent>
-              <validateDataStore>false</validateDataStore>
-              <!-- Don't want datastore... -->
-              <validateDataSet>false</validateDataSet>
-              <!-- ...nor dataset -->
-              <validateActions>true</validateActions>
-              <validateDocumentation>true</validateDocumentation>
-              <validateWording>true</validateWording>
-              <validateLayout>true</validateLayout>
-              <validateOptionNames>true</validateOptionNames>
-              <validateOutputConnection>true</validateOutputConnection>
-              <validateLocalConfiguration>true</validateLocalConfiguration>
-              <validateNoFinalOption>true</validateNoFinalOption>
-              <validateExceptions>true</validateExceptions>
-              <failOnValidateExceptions>true</failOnValidateExceptions>
-              <validateRecord>true</validateRecord>
-              <validateSchema>true</validateSchema>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-jar-plugin</artifactId>
-        <configuration>
-          <archive>
-            <manifest>
-              <addClasspath>true</addClasspath>
-              <mainClass>org.talend.sdk.component.sample.feature.supporterror.Cli</mainClass>
-            </manifest>
-          </archive>
-        </configuration>
-      </plugin>
-      <plugin>
-        <artifactId>maven-shade-plugin</artifactId>
-        <version>2.1</version>
-        <executions>
-          <execution>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <transformers>
-                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>org.talend.sdk.component.sample.feature.supporterror.Cli</mainClass>
-                </transformer>
-              </transformers>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>exec-maven-plugin</artifactId>
-        <version>3.0.0</version>
-        <configuration>
-          <mainClass>org.talend.sdk.component.sample.feature.supporterror.Cli</mainClass>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
