--- sample-parent/sample-features/supporterror/src/main/java/org/talend/sdk/component/sample/feature/supporterror/package-info.java
@@ -1,21 +1,0 @@
-/**
- * Copyright (C) 2006-2025 Talend Inc. - www.talend.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@Components(family = "sampleRecordWithEntriesInError", categories = "sample")
-@Icon(value = Icon.IconType.CUSTOM, custom = "icon")
-package org.talend.sdk.component.sample.feature.supporterror;
-
-import org.talend.sdk.component.api.component.Components;
-import org.talend.sdk.component.api.component.Icon;
